move submit login and sign-up functions to requests.js
